+++ src/app/miru/miruHomeQueries.ts
@@ -4,32 +4,58 @@ export async function GetRatedAnime() {
     const query = 
     `
     query {
-        animeByCategory(category: "score", count: 5) {
-            id,
-            title,
-            slug,
-            coverImgUrl
+        miru {
+            animes (
+                sort: {
+                    category: "score",
+                    direction: "desc"
+                },
+                pagination: {
+                    perPage: 5,
+                    targetPage: 1
+                }
+                ) {
+                results {
+                    id,
+                    title,
+                    slug,
+                    coverImageUrl
+                }
+            }
         }
     }
     `
 
     const res = await arcadiaAPI.GraphQL<any>(query)
-    return res.data.animeByCategory
+    return res.data.miru.animes.results
 }
 
 export async function GetPopularAnime() {
     const query = 
     `
     query {
-        animeByCategory(category: "users", count: 5) {
-            id,
-            title,
-            slug,
-            coverImgUrl
+        miru {
+            animes (
+                sort: {
+                    category: "users",
+                    direction: "desc"
+                },
+                pagination: {
+                    perPage: 5,
+                    targetPage: 1
+                }
+                ) {
+                results {
+                    id,
+                    title,
+                    slug,
+                    coverImageUrl
+                }
+            }
         }
     }
     `
 
     const res = await arcadiaAPI.GraphQL<any>(query)
-    return res.data.animeByCategory
+    return res.data.miru.animes.results
 }